Improve error reporting.
The name of option is now shown in most cases when it's feasible, and
clarify of the error messages has been improved throughout.

Patch from Leo Goodstadt.


git-svn-id: http://svn.boost.org/svn/boost/trunk@77827 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/program_options/example/first.cpp
@@ -20,7 +20,7 @@ int main(int ac, char* av[])
         po::options_description desc("Allowed options");
         desc.add_options()
             ("help", "produce help message")
-            ("compression", po::value<int>(), "set compression level")
+            ("compression", po::value<double>(), "set compression level")
         ;
 
         po::variables_map vm;        
@@ -34,7 +34,7 @@ int main(int ac, char* av[])
 
         if (vm.count("compression")) {
             cout << "Compression level was set to " 
-                 << vm["compression"].as<int>() << ".\n";
+                 << vm["compression"].as<double>() << ".\n";
         } else {
             cout << "Compression level was not set.\n";
         }
